% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctseg.R, R/ctseg_full.R
\name{predict_ctseg}
\alias{predict_ctseg}
\alias{register_ctseg}
\alias{download_ctseg_model}
\alias{download_ctbet_model}
\alias{load_ctbet_model}
\alias{load_ctseg_model}
\title{Predict Head Segmentation of CT images}
\usage{
predict_ctseg(
  image,
  register = FALSE,
  verbose = TRUE,
  ...,
  outdir = NULL,
  type = c("NPH", "Normal_Only")
)

register_ctseg(image, mask = NULL, verbose = TRUE, ...)

download_ctseg_model(outdir = NULL, type = c("NPH", "Normal_Only"))

download_ctbet_model(
  outdir = NULL,
  type = c("unet_CT_SS_20171114_170726.h5", "unet_CT_SS_3D_201843_163521.h5")
)

load_ctbet_model(
  outdir = NULL,
  type = c("unet_CT_SS_20171114_170726.h5", "unet_CT_SS_3D_201843_163521.h5")
)

load_ctseg_model(outdir = NULL, type = c("NPH", "Normal_Only"))
}
\arguments{
\item{image}{image to segment using `HeadCTSegmentation` model}

\item{verbose}{print diagnostic messages}

\item{outdir}{Output directory for `CTSeg` model}
}
\value{
A list of the output images and predictions.
}
\description{
Predict Head Segmentation of CT images

CTSeg Model
}
\note{
\url{https://github.com/jasonccai/HeadCTSegmentation/}
}
\examples{
url = paste0("https://github.com/jasonccai/HeadCTSegmentation/",
"raw/master/image_data_predict/1.nii.gz")
image = tempfile(fileext = ".nii.gz")
utils::download.file(url, destfile = image, quiet = FALSE)
out = predict_ctseg(image)
\dontrun{
reg_out = predict_ctseg(image, register = TRUE)
}
\donttest{
destfile = file.path(tempdir(), "01.tar.xz")
dl = download.file(
  "https://archive.data.jhu.edu/api/access/datafile/1311?gbrecs=true",
  destfile = destfile)
res = untar(tarfile = destfile, exdir = tempdir())
fname = file.path(tempdir(), "01", "BRAIN_1_Anonymized.nii.gz")
mask = file.path(tempdir(), "01", "BRAIN_1_Anonymized_Mask.nii.gz")
tdir = tempfile()
dir.create(tdir)
download_ctseg_model(outdir = tdir)
mod = load_ctseg_model(outdir = tdir)
predict_ctseg(fname, mask = mask, outdir = tdir)
}
}
