% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepbleed.R
\name{download_ctseg_model}
\alias{download_ctseg_model}
\title{DeepBleed Model}
\usage{
download_ctseg_model(outdir = NULL, type = c("NPH", "Normal_Only"))
}
\arguments{
\item{outdir}{Output directory for `DeepBleed` model}
}
\value{
A list of the output images and predictions.
}
\description{
DeepBleed Model
}
\note{
\url{https://github.com/muschellij2/deepbleed}
}
\examples{
\donttest{
destfile = file.path(tempdir(), "01.tar.xz")
dl = download.file(
  "https://archive.data.jhu.edu/api/access/datafile/1311?gbrecs=true",
  destfile = destfile)
res = untar(tarfile = destfile, exdir = tempdir())
fname = file.path(tempdir(), "01", "BRAIN_1_Anonymized.nii.gz")
mask = file.path(tempdir(), "01", "BRAIN_1_Anonymized_Mask.nii.gz")
tdir = tempfile()
dir.create(tdir)
download_ctseg_model(outdir = tdir)
mod = load_ctseg_model(outdir = tdir)
predict_ctseg(fname, mask = mask, outdir = tdir)
}
}
