% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctbet.R
\name{predict_ctbet}
\alias{predict_ctbet}
\title{Segment Brain from Scan with CTBET CNN}
\usage{
predict_ctbet(
  image,
  weight_file = NULL,
  dimension = 2L,
  data_augmentation = TRUE
)
}
\arguments{
\item{image}{Image or set of images to segment}

\item{weight_file}{Model weight file, see \code{\link{download_ctbet_model}}}

\item{data_augmentation}{Should data augmentation be done before
prediction?}
}
\value{
A list of \code{nifti}
}
\description{
Segment Brain from Scan with CTBET CNN
}
